Moved remove credential function to http utils
