+++ instrumentation/opentelemetry-instrumentation-asgi/setup.cfg
@@ -41,6 +41,7 @@ install_requires =
     opentelemetry-api == 1.4.0.dev0
     opentelemetry-semantic-conventions == 0.23.dev0
     opentelemetry-instrumentation == 0.23.dev0
+    opentelemetry-util-http == 0.23.dev0
 
 [options.extras_require]
 test =
